requirements.txt
asgiref==3.3.1
boto==2.49.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
